Simplify modal component
